requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-mongodb-backend==5.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.3.7
dnspython==2.8.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.4
mongoengine==0.29.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pur==7.3.3
PyJWT==2.10.1
pymongo==4.15.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
